Small refactors
