+++ plugins/grid-bookmark/src/GridBookmarkWidget/index.ts
@@ -1,3 +1,19 @@
+import { lazy } from 'react'
 import { ConfigurationSchema } from '@jbrowse/core/configuration'
-export { default as stateModelFactory } from './model'
-export const configSchema = ConfigurationSchema('GridBookmarkWidget', {})
+import { WidgetType } from '@jbrowse/core/pluggableElementTypes'
+import stateModelFactory from './model'
+import PluginManager from '@jbrowse/core/PluginManager'
+
+const configSchema = ConfigurationSchema('GridBookmarkWidget', {})
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addWidgetType(() => {
+    return new WidgetType({
+      name: 'GridBookmarkWidget',
+      heading: 'Bookmarked regions',
+      configSchema,
+      stateModel: stateModelFactory(pluginManager),
+      ReactComponent: lazy(() => import('./components/GridBookmarkWidget')),
+    })
+  })
+}
